tests: split tests and allow tests to be run without plot requirements
